+++ amazon-chime-sdk/src/main/java/com/amazonaws/services/chime/sdk/meetings/analytics/DefaultMeetingStatsCollector.kt
@@ -15,6 +15,10 @@ class DefaultMeetingStatsCollector(
     private var meetingStartConnectingTimeMs: Long = 0L
     // The time meeting has started
     private var meetingStartTimeMs: Long = 0L
+    // The time that meeting starts reconnecting
+    private var meetingStartReconnectingTimeMs: Long = 0L
+    // The time that meeting starts reconnecting
+    private var meetingReconnectedTimeMs: Long = 0L
     // The number of attempts to reconnect to the meeting
     private var retryCount: Int = 0
     // The number of poor connection occurrences
@@ -44,21 +48,37 @@ class DefaultMeetingStatsCollector(
         meetingStartTimeMs = Calendar.getInstance().timeInMillis
     }
 
+    override fun updateMeetingStartReconnectingTimeMs() {
+        meetingStartReconnectingTimeMs = Calendar.getInstance().timeInMillis
+        meetingReconnectedTimeMs = 0L
+    }
+
+    override fun updateMeetingReconnectedTimeMs() {
+        meetingReconnectedTimeMs = Calendar.getInstance().timeInMillis
+    }
+
     override fun resetMeetingStats() {
         meetingStartConnectingTimeMs = 0L
         meetingStartTimeMs = 0L
         retryCount = 0
         poorConnectionCount = 0
         maxVideoTileCount = 0
+        meetingStartReconnectingTimeMs = 0L
+        meetingReconnectedTimeMs = 0L
     }
 
     override fun getMeetingStatsEventAttributes(): EventAttributes {
+        val meetingReconnectDurationMs =
+            if (meetingStartReconnectingTimeMs == 0L || meetingReconnectedTimeMs == 0L || meetingReconnectedTimeMs < meetingStartReconnectingTimeMs) 0L
+            else meetingReconnectedTimeMs - meetingStartReconnectingTimeMs
+
         return mutableMapOf(
             EventAttributeName.maxVideoTileCount to maxVideoTileCount,
             EventAttributeName.retryCount to retryCount,
             EventAttributeName.poorConnectionCount to poorConnectionCount,
             EventAttributeName.meetingDurationMs to if (meetingStartTimeMs == 0L) 0L else Calendar.getInstance().timeInMillis - meetingStartTimeMs,
-            EventAttributeName.meetingStartDurationMs to if (meetingStartTimeMs == 0L) 0L else meetingStartTimeMs - meetingStartConnectingTimeMs
+            EventAttributeName.meetingStartDurationMs to if (meetingStartTimeMs == 0L) 0L else meetingStartTimeMs - meetingStartConnectingTimeMs,
+            EventAttributeName.meetingReconnectDurationMs to meetingReconnectDurationMs
         )
     }