+++ amazon-chime-sdk/src/main/java/com/amazonaws/services/chime/sdk/meetings/device/DefaultDeviceController.kt
@@ -18,6 +18,8 @@ import android.os.Build
 import androidx.annotation.RequiresApi
 import androidx.annotation.VisibleForTesting
 import com.amazonaws.services.chime.sdk.meetings.analytics.EventAnalyticsController
+import com.amazonaws.services.chime.sdk.meetings.analytics.EventAttributeName
+import com.amazonaws.services.chime.sdk.meetings.analytics.EventName
 import com.amazonaws.services.chime.sdk.meetings.analytics.MeetingHistoryEventName
 import com.amazonaws.services.chime.sdk.meetings.internal.audio.AudioClientController
 import com.amazonaws.services.chime.sdk.meetings.internal.audio.AudioClientState
@@ -25,13 +27,17 @@ import com.amazonaws.services.chime.sdk.meetings.internal.audio.DefaultAudioClie
 import com.amazonaws.services.chime.sdk.meetings.internal.utils.ConcurrentSet
 import com.amazonaws.services.chime.sdk.meetings.internal.utils.ObserverUtils
 import com.amazonaws.services.chime.sdk.meetings.internal.video.VideoClientController
+import com.amazonaws.services.chime.sdk.meetings.utils.MediaError
+import com.amazonaws.services.chime.sdk.meetings.utils.logger.Logger
 import com.xodee.client.audio.audioclient.AudioClient
+import kotlin.Any
 
 class DefaultDeviceController(
     private val context: Context,
     private val audioClientController: AudioClientController,
     private val videoClientController: VideoClientController,
     private val eventAnalyticsController: EventAnalyticsController,
+    private val logger: Logger,
     private val audioManager: AudioManager = context.getSystemService(Context.AUDIO_SERVICE) as AudioManager,
     private val buildVersion: Int = Build.VERSION.SDK_INT
 ) : DeviceController {
@@ -43,6 +49,8 @@ class DefaultDeviceController(
     private var receiver: BroadcastReceiver? = null
 
     private var audioDeviceCallback: AudioDeviceCallback? = null
+
+    private val TAG = "DefaultDeviceController"
 
     init {
         @SuppressLint("NewApi")
@@ -117,7 +125,13 @@ class DefaultDeviceController(
                     )
                 )
             }
-
+            if (audioDevices.isEmpty()) {
+                val attributes = mutableMapOf<EventAttributeName, Any>(
+                    EventAttributeName.audioInputErrorMessage to MediaError.NoAudioDevices
+                )
+                eventAnalyticsController.publishEvent(EventName.audioInputFailed, attributes)
+                logger.error(TAG, "List audio devices failed: ${MediaError.NoAudioDevices}")
+            }
             // It doesn't look like Android can switch between two wired connection, so we'll assume WIRED_HEADSET
             // is where audio is routed.
             if (wiredDeviceCount > 1) audioDevices.removeIf { it.type == MediaDeviceType.AUDIO_USB_HEADSET }