+++ amazon-chime-sdk/src/main/java/com/amazonaws/services/chime/sdk/meetings/session/DefaultMeetingSession.kt
@@ -150,7 +150,8 @@ class DefaultMeetingSession @JvmOverloads constructor(
                 context,
                 audioClientController,
                 videoClientController,
-                eventAnalyticsController
+                eventAnalyticsController,
+                logger
             )
 
         val realtimeController =