+++ amazon-chime-sdk/src/test/java/com/amazonaws/services/chime/sdk/meetings/analytics/DefaultEventAnalyticsControllerTest.kt
@@ -6,15 +6,19 @@
 package com.amazonaws.services.chime.sdk.meetings.analytics
 
 import com.amazonaws.services.chime.sdk.meetings.ingestion.EventReporter
+import com.amazonaws.services.chime.sdk.meetings.internal.ingestion.SDKEvent
 import com.amazonaws.services.chime.sdk.meetings.internal.utils.DeviceUtils
 import com.amazonaws.services.chime.sdk.meetings.session.MeetingSessionConfiguration
 import com.amazonaws.services.chime.sdk.meetings.session.MeetingSessionCredentials
 import com.amazonaws.services.chime.sdk.meetings.utils.logger.Logger
 import io.mockk.MockKAnnotations
+import io.mockk.Runs
 import io.mockk.every
 import io.mockk.impl.annotations.MockK
+import io.mockk.just
 import io.mockk.mockkObject
 import io.mockk.mockkStatic
+import io.mockk.slot
 import io.mockk.spyk
 import io.mockk.unmockkObject
 import io.mockk.verify
@@ -28,6 +32,8 @@ import kotlinx.coroutines.test.setMain
 import org.junit.After
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertFalse
+import org.junit.Assert.assertNotNull
+import org.junit.Assert.assertNull
 import org.junit.Assert.assertTrue
 import org.junit.Before
 import org.junit.Test
@@ -50,12 +56,16 @@ class DefaultEventAnalyticsControllerTest {
     @MockK
     private lateinit var mockMeetingStatsCollector: MeetingStatsCollector
 
-    private val meetingDurationAttribute: EventAttributes =
-        mutableMapOf(EventAttributeName.meetingDurationMs to 1000L)
+    private val meetingAttributes: EventAttributes = mutableMapOf()
+    private val mockMeetingDurationMs = 1000L
+    private val mockMeetingReconnectDurationMs = 500L
 
     @Before
     fun setup() {
         mockkStatic(Calendar::class)
+
+        meetingAttributes[EventAttributeName.meetingDurationMs] = mockMeetingDurationMs
+        meetingAttributes[EventAttributeName.meetingReconnectDurationMs] = mockMeetingReconnectDurationMs
 
         MockKAnnotations.init(this, relaxUnitFun = true)
         testEventAnalyticsController =
@@ -65,7 +75,7 @@ class DefaultEventAnalyticsControllerTest {
                 mockMeetingStatsCollector,
                 eventReporter
             )
-        every { mockMeetingStatsCollector.getMeetingStatsEventAttributes() } returns meetingDurationAttribute
+        every { mockMeetingStatsCollector.getMeetingStatsEventAttributes() } returns meetingAttributes
         every { mockMeetingStatsCollector.getMeetingHistory() } returns emptyList()
         every { mockMeetingSessionConfiguration.credentials } returns MeetingSessionCredentials(
             "attendeeId",
@@ -123,7 +133,32 @@ class DefaultEventAnalyticsControllerTest {
     fun `publishEvent should invoked EventReporter's report`() {
         testEventAnalyticsController.publishEvent(EventName.meetingFailed)
 
+        every { eventReporter.report(any()) }
+
         verify(exactly = 1) { eventReporter.report(any()) }
+    }
+
+    @Test
+    fun `publishEvent should remove meetingReconnectedDurationMs if not meetingReconnected event`() {
+        val slot = slot<SDKEvent>()
+        every { eventReporter.report(capture(slot)) } just Runs
+        testEventAnalyticsController.publishEvent(EventName.meetingFailed)
+
+        verify(exactly = 1) { eventReporter.report(any()) }
+        assertNotNull(slot.captured.eventAttributes[EventAttributeName.meetingDurationMs.name])
+        assertNull(slot.captured.eventAttributes[EventAttributeName.meetingReconnectDurationMs.name])
+    }
+
+    @Test
+    fun `publishEvent should contain meetingReconnectedDurationMs if meetingReconnected event`() {
+        val slot = slot<SDKEvent>()
+        every { eventReporter.report(capture(slot)) } just Runs
+
+        testEventAnalyticsController.publishEvent(EventName.meetingReconnected)
+
+        verify(exactly = 1) { eventReporter.report(any()) }
+        assertNotNull(slot.captured.eventAttributes[EventAttributeName.meetingDurationMs.name])
+        assertNotNull(slot.captured.eventAttributes[EventAttributeName.meetingReconnectDurationMs.name])
     }
 
     @Test