+++ amazon-chime-sdk/src/test/java/com/amazonaws/services/chime/sdk/meetings/analytics/DefaultMeetingStatsCollectorTest.kt
@@ -47,17 +47,21 @@ class DefaultMeetingStatsCollectorTest {
         testMeetingStatsCollector.incrementPoorConnectionCount()
         testMeetingStatsCollector.incrementRetryCount()
         testMeetingStatsCollector.updateMaxVideoTile(5)
-
+        testMeetingStatsCollector.updateMeetingStartReconnectingTimeMs()
+        Thread.sleep(1000)
+        testMeetingStatsCollector.updateMeetingReconnectedTimeMs()
         testMeetingStatsCollector.resetMeetingStats()
 
         val meetingDurationMs = testMeetingStatsCollector.getMeetingStatsEventAttributes()[EventAttributeName.meetingDurationMs]
         val retryCount = testMeetingStatsCollector.getMeetingStatsEventAttributes()[EventAttributeName.retryCount]
         val poorConnectionCount = testMeetingStatsCollector.getMeetingStatsEventAttributes()[EventAttributeName.poorConnectionCount]
         val maxVideoTileCount = testMeetingStatsCollector.getMeetingStatsEventAttributes()[EventAttributeName.maxVideoTileCount]
+        val meetingReconnectDurationMs = testMeetingStatsCollector.getMeetingStatsEventAttributes()[EventAttributeName.meetingReconnectDurationMs]
         assertEquals(0L, meetingDurationMs)
         assertEquals(0, retryCount)
         assertEquals(0, poorConnectionCount)
         assertEquals(0, maxVideoTileCount)
+        assertEquals(0L, meetingReconnectDurationMs)
     }
 
     @Test
@@ -67,5 +71,37 @@ class DefaultMeetingStatsCollectorTest {
         testMeetingStatsCollector.addMeetingHistoryEvent(MeetingHistoryEventName.meetingStartSucceeded, 100)
 
         assertEquals(1, testMeetingStatsCollector.getMeetingHistory().size)
+    }
+
+    @Test
+    fun `getMeetingStatsEventAttributes should return 0 if meetingStartReconnectingTimeMs is not set`() {
+        val meetingReconnectDurationMs = testMeetingStatsCollector.getMeetingStatsEventAttributes()[EventAttributeName.meetingReconnectDurationMs]
+        assertEquals(0L, meetingReconnectDurationMs)
+    }
+
+    @Test
+    fun `getMeetingStatsEventAttributes should return 0 if meetingReconnectedTimeMs is not set`() {
+        testMeetingStatsCollector.updateMeetingReconnectedTimeMs()
+        val meetingReconnectDurationMs = testMeetingStatsCollector.getMeetingStatsEventAttributes()[EventAttributeName.meetingReconnectDurationMs]
+        assertEquals(0L, meetingReconnectDurationMs)
+    }
+
+    @Test
+    fun `getMeetingStatsEventAttributes should return 0 if meetingReconnectedTimeMs is eailer than meetingStartReconnectingTimeMs`() {
+        testMeetingStatsCollector.updateMeetingReconnectedTimeMs()
+        testMeetingStatsCollector.updateMeetingStartReconnectingTimeMs()
+
+        val meetingReconnectDurationMs = testMeetingStatsCollector.getMeetingStatsEventAttributes()[EventAttributeName.meetingReconnectDurationMs]
+        assertEquals(0L, meetingReconnectDurationMs)
+    }
+
+    @Test
+    fun `getMeetingStatsEventAttributes should return the duration if meetingStartReconnectingTimeMs and meetingReconnectedTimeMs are updated properly`() {
+        testMeetingStatsCollector.updateMeetingStartReconnectingTimeMs()
+        Thread.sleep(1000)
+        testMeetingStatsCollector.updateMeetingReconnectedTimeMs()
+
+        val meetingReconnectDurationMs = testMeetingStatsCollector.getMeetingStatsEventAttributes()[EventAttributeName.meetingReconnectDurationMs] as Long
+        assert(meetingReconnectDurationMs in 950L..1050L)
     }
 }