+++ amazon-chime-sdk/src/test/java/com/amazonaws/services/chime/sdk/meetings/device/DefaultDeviceControllerTest.kt
@@ -11,11 +11,15 @@ import android.media.AudioDeviceInfo
 import android.media.AudioManager
 import android.media.AudioRecordingConfiguration
 import com.amazonaws.services.chime.sdk.meetings.analytics.EventAnalyticsController
+import com.amazonaws.services.chime.sdk.meetings.analytics.EventAttributeName
+import com.amazonaws.services.chime.sdk.meetings.analytics.EventName
 import com.amazonaws.services.chime.sdk.meetings.analytics.MeetingHistoryEventName
 import com.amazonaws.services.chime.sdk.meetings.internal.audio.AudioClientController
 import com.amazonaws.services.chime.sdk.meetings.internal.audio.AudioClientState
 import com.amazonaws.services.chime.sdk.meetings.internal.audio.DefaultAudioClientController
 import com.amazonaws.services.chime.sdk.meetings.internal.video.VideoClientController
+import com.amazonaws.services.chime.sdk.meetings.utils.MediaError
+import com.amazonaws.services.chime.sdk.meetings.utils.logger.Logger
 import com.xodee.client.audio.audioclient.AudioClient
 import io.mockk.MockKAnnotations
 import io.mockk.every
@@ -23,6 +27,7 @@ import io.mockk.impl.annotations.MockK
 import io.mockk.mockkClass
 import io.mockk.mockkStatic
 import io.mockk.verify
+import kotlin.Any
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.TestCoroutineDispatcher
@@ -76,6 +81,9 @@ class DefaultDeviceControllerTest {
     @MockK
     private lateinit var deviceChangeObserver: DeviceChangeObserver
 
+    @MockK
+    private lateinit var mockLogger: Logger
+
     private lateinit var deviceController: DefaultDeviceController
 
     private val testDispatcher = TestCoroutineDispatcher()
@@ -87,6 +95,7 @@ class DefaultDeviceControllerTest {
             audioClientController,
             videoClientController,
             eventAnalyticsController,
+            mockLogger,
             audioManager,
             24
         )
@@ -101,6 +110,7 @@ class DefaultDeviceControllerTest {
             audioClientController,
             videoClientController,
             eventAnalyticsController,
+            mockLogger,
             audioManager,
             21
         )
@@ -204,6 +214,19 @@ class DefaultDeviceControllerTest {
                         it.label == "my bluetooth headphone (Bluetooth)"
             )
         }
+    }
+
+    @Test
+    fun `listAudioDevices should public audioInputFailed event when no device available`() {
+        setupForNewAPILevel()
+        every { audioManager.getDevices(AudioManager.GET_DEVICES_OUTPUTS) } returns arrayOf()
+
+        deviceController.listAudioDevices()
+
+        val attributes = mutableMapOf<EventAttributeName, Any>(
+            EventAttributeName.audioInputErrorMessage to MediaError.NoAudioDevices
+        )
+        verify(exactly = 1) { eventAnalyticsController.publishEvent(EventName.audioInputFailed, attributes) }
     }
 
     @Test