+++ amazon-chime-sdk/src/test/java/com/amazonaws/services/chime/sdk/meetings/internal/audio/DefaultAudioClientControllerTest.kt
@@ -22,6 +22,7 @@ import com.amazonaws.services.chime.sdk.meetings.audiovideo.audio.AudioRecording
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.audio.AudioStreamType
 import com.amazonaws.services.chime.sdk.meetings.internal.utils.AppInfoUtil
 import com.amazonaws.services.chime.sdk.meetings.session.MeetingSessionStatusCode
+import com.amazonaws.services.chime.sdk.meetings.utils.MediaError
 import com.amazonaws.services.chime.sdk.meetings.utils.logger.Logger
 import com.xodee.client.audio.audioclient.AppInfo
 import com.xodee.client.audio.audioclient.AudioClient
@@ -205,6 +206,11 @@ class DefaultAudioClientControllerTest {
         val testOutput: Boolean = audioClientController.setRoute(testNewRoute)
 
         assertFalse(testOutput)
+
+        val attributes = mutableMapOf<EventAttributeName, Any>(
+            EventAttributeName.audioInputErrorMessage to MediaError.FailedToSetRoute
+        )
+        verify { mockEventAnalyticsController.publishEvent(EventName.audioInputFailed, attributes) }
     }
 
     @Test